requirements.txt
black
flake8
isort
pycryptodome
whirlpool
